implemented the photos query and refactoring
